added multiple events production and consumption for communication

+++ common-lib/src/main/java/com/vsnt/common_lib/dtos/events/asset/transcoding/AssetTranscodingProgressPayload.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class AssetTranscodingProgressPayload {
-    private int progressPercentage;
+    private double progressPercentage;
 }
